+++ WEB-INF/src/org/digijava/module/aim/form/IPAContractForm.java
@@ -1,5 +1,6 @@
 package org.digijava.module.aim.form;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -93,8 +94,8 @@ public class IPAContractForm extends ActionForm  {
 	private String totalPrivateContribAmount;
 	private String totalPrivateContribAmountDate;
 
-	private Double totalDisbursements;
-    private Double executionRate;
+	private BigDecimal totalDisbursements;
+    private BigDecimal executionRate;
 	
 	private List disbursements;
     private List<AmpOrganisation> organisations;
@@ -379,19 +380,19 @@ public class IPAContractForm extends ActionForm  {
 		this.dibusrsementsGlobalCurrency = dibusrsementsGlobalCurrency;
 	}
 
-	public Double getTotalDisbursements() {
+	public BigDecimal getTotalDisbursements() {
 		return totalDisbursements;
 	}
 
-	public void setTotalDisbursements(Double totalDisbursements) {
+	public void setTotalDisbursements(BigDecimal totalDisbursements) {
 		this.totalDisbursements = totalDisbursements;
 	}
 
-	public Double getExecutionRate() {
+	public BigDecimal getExecutionRate() {
 		return executionRate;
 	}
 
-	public void setExecutionRate(Double executionRate) {
+	public void setExecutionRate(BigDecimal executionRate) {
 		this.executionRate = executionRate;
 	}
