+++ WEB-INF/src/org/digijava/module/aim/helper/Components.java
@@ -5,6 +5,7 @@
 package org.digijava.module.aim.helper;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Map;
 import java.util.SortedMap;
@@ -26,7 +27,7 @@ public class Components<T> implements Comparable , Serializable{
 	private String url;
 	private Long  type_Id;
 	
-	private SortedMap<Integer,Map<String,Double>> financeByYearInfo;
+	private SortedMap<Integer,Map<String,BigDecimal>> financeByYearInfo;
 	
 	public Components() {}
 	
@@ -195,11 +196,11 @@ public class Components<T> implements Comparable , Serializable{
 		return url;
 	}
 
-	public void setFinanceByYearInfo(SortedMap<Integer,Map<String,Double>> financeByYearInfo) {
+	public void setFinanceByYearInfo(SortedMap<Integer,Map<String,BigDecimal>> financeByYearInfo) {
 		this.financeByYearInfo = financeByYearInfo;
 	}
 
-	public SortedMap<Integer,Map<String,Double>> getFinanceByYearInfo() {
+	public SortedMap<Integer,Map<String,BigDecimal>> getFinanceByYearInfo() {
 		return financeByYearInfo;
 	}
